+++ libs/maps/include/mrpt/maps/CHeightGridMap2D_MRF.h
@@ -94,7 +94,7 @@ class CHeightGridMap2D_MRF : public CRandomFieldGridMap2D,
 		const mrpt::poses::CPose3D* robotPose = nullptr) override;
 	double internal_computeObservationLikelihood(
 		const mrpt::obs::CObservation& obs,
-		const mrpt::poses::CPose3D& takenFrom) override;
+		const mrpt::poses::CPose3D& takenFrom) const override;
 
 	MAP_DEFINITION_START(CHeightGridMap2D_MRF)
 	/** Runs map estimation at start up (Default:true) */