+++ libs/maps/src/maps/CHeightGridMap2D_MRF.cpp
@@ -152,7 +152,7 @@ bool CHeightGridMap2D_MRF::internal_insertObservation(
   ---------------------------------------------------------------*/
 double CHeightGridMap2D_MRF::internal_computeObservationLikelihood(
 	[[maybe_unused]] const CObservation& obs,
-	[[maybe_unused]] const CPose3D& takenFrom)
+	[[maybe_unused]] const CPose3D& takenFrom) const
 {
 	THROW_EXCEPTION("Not implemented yet!");
 }