+++ libs/slam/include/mrpt/slam/TMonteCarloLocalizationParams.h
@@ -19,29 +19,18 @@ namespace mrpt::slam
  */
 struct TMonteCarloLocalizationParams
 {
-	/** Default settings method.
-	 */
-	TMonteCarloLocalizationParams();
-
-	/** Copy constructor: take care of knowing what you do, since this copies
-	 * pointers. */
-	TMonteCarloLocalizationParams(const TMonteCarloLocalizationParams& o);
-
-	/** Copy operator: take care of knowing what you do, since this copies
-	 * pointers. */
-	TMonteCarloLocalizationParams& operator=(
-		const TMonteCarloLocalizationParams& o);
+	TMonteCarloLocalizationParams() = default;
 
 	/** [update stage] Must be set to a metric map used to estimate the
 	 * likelihood of observations
 	 */
-	mrpt::maps::CMetricMap* metricMap{nullptr};
+	mrpt::maps::CMetricMap::ConstPtr metricMap;
 
 	/** [update stage] Alternative way (if metricMap==nullptr): A metric map is
 	 * supplied for each particle: There must be the same maps here as pose
 	 * m_particles.
 	 */
-	mrpt::maps::TMetricMapList metricMaps;
+	std::vector<mrpt::maps::CMetricMap::ConstPtr> metricMaps;
 
 	/** Parameters for dynamic sample size, KLD method. */
 	TKLDParams KLD_params;