+++ doc/source/doxygen-docs/changelog.md
@@ -16,8 +16,12 @@
     - New CMake build flag `MRPT_EXCEPTIONS_WITH_CALL_STACK` to optionally disable reporting call stacks upon exceptions.
   - \ref mrpt_math_grp
     - New class mrpt::math::TBoundingBox
+    - \ref mrpt_maps_grp
+      - Const correctness fixed in all mrpt::maps::CMetricMap classes.
   - \ref mrpt_opengl_grp
     - mrpt::opengl::CFrustum() new constructor from mrpt::img::TCamera()
+  - \ref mrpt_slam_grp
+    - mrpt::slam::TMonteCarloLocalizationParams map parameters are now shared pointers instead of plain pointers for safer code.
 - BUG FIXES:
   - Log `*_THROTTLE_*` macros (e.g. MRPT_LOG_THROTTLE_DEBUG) did not report the message the first time they were called, which seems a safer behavior.
   - Reverted changed behavior: mrpt::config::CConfigFile did not throw if a non-existing file was passed to its constructor, but it throws in MRPT 2.1.{0-7}.